+++ demo_modules/slideshow/client_interfaces.ts
@@ -1,5 +1,7 @@
+/// <reference lib="dom" />
+
 import { Rectangle } from "../../lib/math/rectangle.ts";
-import { Content, ContentId } from "./interfaces.ts";
+import { ContentId } from "./interfaces.ts";
 
 export interface ClientLoadStrategy {
   /**
@@ -13,4 +15,19 @@ export interface ClientLoadStrategy {
 export interface ClientDisplayStrategy {
   /** Update the surface with the content. */
   draw(time: number, delta: number): void;
+}
+
+export interface Content {
+  /** The type of the content. */
+  readonly type: "video" | "image";
+  /** Returns the width of the content. */
+  readonly width: number;
+  /** Returns the height of the content. */
+  readonly height: number;
+  /** The 'size' of the content, in terms of bytes. */
+  readonly size: number;
+  /** Some Element that refers to the content itself. */
+  element: HTMLElement;
+  /** Some pieces of content have a draw function. */
+  draw?: (startTime: number, time: number, delta: number) => void;
 }
