+++ demo_modules/slideshow/interfaces.ts
@@ -30,21 +30,6 @@ export interface ContentId {
   id: string;
 }
 
-export interface Content {
-  /** The type of the content. */
-  readonly type: "video" | "image";
-  /** Returns the width of the content. */
-  readonly width: number;
-  /** Returns the height of the content. */
-  readonly height: number;
-  /** The 'size' of the content, in terms of bytes. */
-  readonly size: number;
-  /** Some Element that refers to the content itself. */
-  element: HTMLElement;
-  /** Some pieces of content have a draw function. */
-  draw?: (startTime: number, time: number, delta: number) => void;
-}
-
 export interface ContentPage {
   paginationToken?: string;
   contentIds: ContentId[];
